requirements.txt
charset-normalizer==3.4.4
customtkinter==5.2.2
darkdetect==0.8.0
ExifRead==3.5.1
packaging==26.0
piexif==1.1.3
pillow==12.1.0
pillow_heif==1.2.0
reportlab==4.4.9
tkinterdnd2==0.4.3
